ex10 marche pas

+++ src/ch/comem/quinteirosm/serie2/ex01_valAbs.java
@@ -7,9 +7,9 @@ public class ex01_valAbs {
     public static void main(String[] args) {
         Scanner userInput = new Scanner(System.in).useLocale(new Locale("fr-CH"));
         System.out.println("Veuillez saisir un nombre :");
-        int reel = userInput.nextInt();
+        double reel = userInput.nextDouble();
 
-        int valeurAbsolue = Math.abs(reel);
+        double valeurAbsolue = Math.abs(reel);
 
         System.out.println("La valeur absolue de "+reel+" est "+valeurAbsolue+" .");
     }
